lint

+++ test/tokens/Weth.sol
@@ -4,15 +4,15 @@ pragma solidity ^0.8.14;
 import { TestERC20 } from "./TestERC20.sol";
 
 contract WETH is TestERC20 {
-    event Deposit(address indexed dst, uint wad);
-    event Withdrawal(address indexed src, uint wad);
+    event Deposit(address indexed dst, uint256 wad);
+    event Withdrawal(address indexed src, uint256 wad);
 
     function deposit() public payable {
         balanceOf[msg.sender] += msg.value;
         emit Deposit(msg.sender, msg.value);
     }
 
-    function withdraw(uint wad) public {
+    function withdraw(uint256 wad) public {
         require(balanceOf[msg.sender] >= wad);
         balanceOf[msg.sender] -= wad;
         address payable sender = payable(msg.sender);
@@ -20,24 +20,26 @@ contract WETH is TestERC20 {
         emit Withdrawal(msg.sender, wad);
     }
 
-    function approve(address guy, uint wad) public override returns (bool) {
+    function approve(address guy, uint256 wad) public override returns (bool) {
         allowance[msg.sender][guy] = wad;
         emit Approval(msg.sender, guy, wad);
         return true;
     }
 
-    function transfer(address dst, uint wad) public override returns (bool) {
+    function transfer(address dst, uint256 wad) public override returns (bool) {
         return transferFrom(msg.sender, dst, wad);
     }
 
     function transferFrom(
         address src,
         address dst,
-        uint wad
+        uint256 wad
     ) public override returns (bool) {
         require(balanceOf[src] >= wad);
 
-        if (src != msg.sender && allowance[src][msg.sender] != type(uint).max) {
+        if (
+            src != msg.sender && allowance[src][msg.sender] != type(uint256).max
+        ) {
             require(allowance[src][msg.sender] >= wad);
             allowance[src][msg.sender] -= wad;
         }
